requirements.txt
# Voiceback - Voice Agent Dependencies
# Core dependencies for Vapi integration and application functionality

# Vapi integration (using requests for HTTP API calls)
requests>=2.32.2

# Environment management
python-dotenv>=1.0.0

# Testing framework
pytest>=7.4.0
pytest-mock>=3.11.1
pytest-cov==4.1.0
pytest-requests-mock==1.11.0

# Logging
loguru>=0.7.2

# HTTP server for webhooks
flask>=2.3.3
gunicorn>=21.2.0

# JSON handling and validation
pydantic>=2.4.0
jsonschema==4.19.0

# Async support
aiohttp>=3.8.6

# Development dependencies
black>=23.9.1
flake8>=6.1.0
mypy==1.5.1

# Additional utilities
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
setuptools==68.2.0

# Security
urllib3==2.0.4
certifi==2023.7.22 